refactor: changed some actions to service code, simplified getting email in rejection
