+++ src/backend/mailing/mailing.service.ts
@@ -1,5 +1,3 @@
-'use server'
-
 import { mailersendClient, MailTemplateId } from '@/lib/mailersendClient'
 import { Recipient } from 'mailersend'
 
@@ -28,6 +26,18 @@ export const sendProfileRejectedEmail = async (
     templateId: templateId,
     config: {
       subject: 'Good DevHunting profile status notification.',
+      fromEmail: 'team@devhunting.co',
+      fromName: 'DevHunting Team',
+    },
+  })
+}
+
+export const sendWelcomeEmail = async (email: string) => {
+  await mailersendClient.sendMail({
+    recipients: [new Recipient(email)],
+    templateId: MailTemplateId.welcomeMail,
+    config: {
+      subject: 'Welcome to Good DevHunting!',
       fromEmail: 'team@devhunting.co',
       fromName: 'DevHunting Team',
     },
